+++ src/main/java/com/github/viniciusmartins/qrcode/controller/QRCodeWithDueDateController.java
@@ -1,9 +1,9 @@
 package com.github.viniciusmartins.qrcode.controller;
 
-import com.github.viniciusmartins.qrcode.controller.interfaces.QRCodeController;
-import com.github.viniciusmartins.qrcode.dto.QRCodeRequest;
-import com.github.viniciusmartins.qrcode.dto.QRCodeResponse;
-import com.github.viniciusmartins.qrcode.service.interfaces.QRCodeService;
+import com.github.viniciusmartins.qrcode.controller.swagger.IQRCodeWithDueDateController;
+import com.github.viniciusmartins.qrcode.dto.QRCodeWithDueDateRequest;
+import com.github.viniciusmartins.qrcode.dto.QRCodeWithDueDateResponse;
+import com.github.viniciusmartins.qrcode.service.QRCodeService;
 import jakarta.validation.Valid;
 import lombok.Cleanup;
 import lombok.RequiredArgsConstructor;
@@ -17,22 +17,22 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping("/api/v1/qrcodes")
+@RequestMapping("/api/v1/qrcodes/with-due-date")
 @RequiredArgsConstructor
 @Slf4j
-public class QRCodeControllerImpl implements QRCodeController {
+public class QRCodeWithDueDateController implements IQRCodeWithDueDateController {
 
-    private static final String LOG_TAG = "QRCODE/TXID";
+    private static final String LOG_TAG = "QRCODE_WITH_DUE_DATE/TXID";
 
     private final QRCodeService qrCodeService;
 
     @Override
     @PostMapping
-    public ResponseEntity<QRCodeResponse> registerQrcode(@RequestBody @Valid QRCodeRequest request) {
+    public ResponseEntity<QRCodeWithDueDateResponse> registerQrcode(@RequestBody @Valid QRCodeWithDueDateRequest request) {
         @Cleanup var mdc = MDC.putCloseable(LOG_TAG, request.txid());
-        log.info("Beginning register QRCode");
-        var response = qrCodeService.register(request);
-        log.info("Ending register QRCode");
+        log.info("Begin register QRCode with due date");
+        var response = qrCodeService.registerWithDueDate(request);
+        log.info("Ending register QRCode with due date");
         return ResponseEntity.status(HttpStatus.CREATED).body(response);
     }
